Added to the copy process and image getwidth/height function now returns an int not unsigned int.

+++ src/Core/Image.h
@@ -29,8 +29,8 @@ class VISI_EXPORT Image
         void AllocateLike(ImageGPU* referanceImage);
 		void Allocate(unsigned int w, unsigned int h, ImageType t);
         void Deallocate();
-        inline unsigned int GetWidth() {return width;}
-		inline unsigned int GetHeight() {return height;}
+        inline int GetWidth() {return (int)width;}
+		inline int GetHeight() {return (int)height;}
         inline ImageType GetType() {return type;}
         inline unsigned char* GetData() {return data;}
         void Copy(ImageGPU* image);
